refactor: moved input.jsx and dropdown.jsx to custom-fields folder for better structure
