+++ src/components/grid-header.jsx
@@ -7,7 +7,7 @@ import { useWindowWidth } from '../hooks/use-window-width';
 import ActionIcon from '../icons/action-icon';
 import { calculateColumnWidth, tryParseWidth } from "../utils/component-utils";
 import ColumnSortIcon from './column-sort-icon';
-import Input from './input';
+import Input from './custom-fields/input';
 
 const GridHeader = ({
     state,
